Assignment8/requirements.txt
Flask==2.3.3
groq==0.4.1
sentence-transformers==2.2.2
PyPDF2==3.0.1
python-dotenv==1.0.0
numpy
python-docx==0.8.11
